docs(readme): Add summarized test cases

+++ src/main/java/lotto/domain/LottoResult.java
@@ -35,7 +35,7 @@ public class LottoResult {
         return totalPrize;
     }
 
-    public static double calculateProfitRate(long totalPrize, int purchaseAmount) {
+    public static double calculateProfitRate(long totalPrize, long purchaseAmount) {
         if (purchaseAmount <= 0) {
             throw new IllegalArgumentException("[ERROR] 구입 금액은 0보다 커야 합니다.");
         }
